+++ app/app.go
@@ -7,8 +7,8 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/gogo/protobuf/proto"
 	"github.com/palomachain/sparrow/attest"
-	chain "github.com/palomachain/sparrow/client"
-	"github.com/palomachain/sparrow/client/paloma"
+	"github.com/palomachain/sparrow/chain"
+	"github.com/palomachain/sparrow/chain/paloma"
 	"github.com/palomachain/sparrow/config"
 	"github.com/palomachain/sparrow/relayer"
 	valsettypes "github.com/palomachain/sparrow/types/paloma/x/valset/types"
