+++ chain/paloma/errors.go
